fix(*): move container.ts into separate folder
having it with other files inside Config causing
a cyclic dependency.
