+++ src/Config/index.ts
@@ -1,4 +1,3 @@
-export { default as container } from './container';
 export * from './CliBuilder';
 export * from './ConfigReader';
 export * from './CliInfo';
